Modify ralloc to use biased lock
We need to store additional metadata in DRAM (process ID, process
count), assuming the caller will provide it when initializing the
heap. Each entrypoint into the API must be protected by the read
side of the lock, to avoid conflicting with the GC.

For now, this doesn't guarantee atomicity of the realloc and
calloc functions: the read lock isn't re-entrant, since `read_unlock`
gives up exclusive access on the first call.